refactor:further adaptations for pc and mobile and minor optimization
