+++ src/App.vue
@@ -2,7 +2,7 @@
 
 <template>
   <router-view></router-view>
-  <footer>@vite-iptv-app</footer>
+  <!-- <footer>@vite-iptv-app</footer> -->
 </template>
 
 <style lang="scss">
