+++ src/pages/TvPlay.vue
@@ -23,60 +23,54 @@
 </template>
 
 <script setup>
+import "video.js/dist/video-js.css";
 import { useRouter } from "vue-router";
 import Videojs from "video.js/dist/video.min.js";
-import { computed, onMounted, onUnmounted, ref } from "vue";
+import { computed, onMounted, onUnmounted } from "vue";
 import { useStore } from "vuex";
 import { message } from "ant-design-vue";
 
 // tv详情模块
-const tvPlayModel = () => {
-  const router = useRouter();
-  const store = useStore();
+const router = useRouter();
+const store = useStore();
 
-  // 回到主页
-  const returnHome = () => {
-    router.push("/");
-  };
+// 回到主页
+const returnHome = () => {
+  router.push("/");
+};
 
-  //播放器与播放状态初始化
-  let myPlyer = null;
+//播放器与播放状态初始化
+let myPlyer = null;
 
-  // 播放频道信息
-  const playerInfo = computed(() => store.state.watching);
+// 播放频道信息
+const playerInfo = computed(() => store.state.watching);
 
-  onMounted(() => {
-    // videojs的第一个参数表示的是，文档中video的id
-    myPlyer = Videojs("videoPlayer", {
-      controls: true,
-      withCredentials: true,
-      playbackRates: [0.5, 1, 1.25, 1.5, 2, 3], // 倍速播放
-      muted: true, //静音模式 解决首次页面加载播放。
-    });
-
-    // 检测播放成功/失败
-    myPlyer.ready(async function () {
-      try {
-        var promise = await myPlyer.play();
-      } catch (error) {
-        message.error("播放失败，源加载出错啦！换一个看吧~");
-        // returnHome();
-      }
-    });
+onMounted(() => {
+  store.commit("deleteSearchResultInfo");
+  // videojs的第一个参数表示的是，文档中video的id
+  myPlyer = Videojs("videoPlayer", {
+    controls: true,
+    withCredentials: true,
+    playbackRates: [0.5, 1, 1.25, 1.5, 2, 3], // 倍速播放
+    muted: true, //静音模式 解决首次页面加载播放。
   });
 
-  onUnmounted(() => {
-    // 离开时将播放器销毁
-    if (myPlyer) {
-      myPlyer.dispose();
+  // 检测播放成功/失败
+  myPlyer.ready(async function () {
+    try {
+      var promise = await myPlyer.play();
+    } catch (error) {
+      message.error("播放失败，源加载出错啦！换一个看吧~");
     }
   });
-  return {
-    returnHome,
-    playerInfo,
-  };
-};
-const { returnHome, playerInfo } = tvPlayModel();
+});
+
+onUnmounted(() => {
+  // 离开时将播放器销毁
+  if (myPlyer) {
+    myPlyer.dispose();
+  }
+});
 </script>
 
 <style lang="scss">
@@ -93,6 +87,7 @@ const { returnHome, playerInfo } = tvPlayModel();
     width: 100px;
     height: 40px;
   }
+
   @media screen and (max-width: 1024px) {
     .vjs-default-skin {
       margin-top: 50px;
