+++ src/router/index.js
@@ -1,9 +1,12 @@
 import { createRouter, createWebHashHistory } from "vue-router";
-import TvPlayPage from "../pages/TvPlayPage.vue";
-import Home from "../pages/Home.vue";
+
+const Home = () => import("../pages/Home/index.vue");
+const TvPlay = () => import("../pages/TvPlay.vue");
+
 const routes = [
   { path: "/", name: "Home", component: Home },
-  { path: "/TvPlayPage", name: "TvPlayPage", component: TvPlayPage },
+  { path: "/TvPlayPage", name: "TvPlayPage", component: TvPlay },
 ];
+
 const router = createRouter({ history: createWebHashHistory(), routes });
 export default router;
